Use puffinn::sort_hashes_pairs_24 in prefixmap
This commit greatly improves the sorting part of indexing. On the
dataset `glove.twitter.27B.25d.txt, using 1Gb of memory for the index,
using `std::sort` requires 6.6 seconds to sort all the prefix maps,
whereas using puffinn::sort_hashes_pairs24 takes only 0.89 seconds
*overall* (on a MacBook pro with dual core Intel I5).

The lion's share of the indexing in this setting is however from
hashing (166 seconds), so this is the next thing to look at for
optimization.